Todo lo referente al radio y mostrar gasolineras en radio, separado en otro script
