+++ gasolineras.js
@@ -2,9 +2,13 @@ export async function buscador_gasolineras(radio, coords) {
     
     let setGasolineras = new Set()
     var fetches=[];
+    
     for (let i=0; i<coords.length;i++){
+        //Seleccionamos la URL en funcion de si se define para una ruta o para una única ubicación
+        var fetchUrl = coords.length > 2 ? "https://api.geoapify.com/v2/places?categories=service.vehicle.fuel&filter=circle:"+String(coords[i][1])+","+String(coords[i][0])+","+String(radio)+"&bias=proximity:"+String(coords[i][1])+","+String(coords[i][0])+"&limit=20&apiKey=5defe68cc4dc4bffb53b9cc477f721f5" 
+        : `https://api.geoapify.com/v2/places?categories=service.vehicle.fuel&filter=circle:${coords[1]},${coords[0]},${radio}&limit=20&apiKey=5defe68cc4dc4bffb53b9cc477f721f5`;
         fetches.push(
-        fetch("https://api.geoapify.com/v2/places?categories=service.vehicle.fuel&filter=circle:"+String(coords[i][1])+","+String(coords[i][0])+","+String(radio)+"&bias=proximity:"+String(coords[i][1])+","+String(coords[i][0])+"&limit=20&apiKey=5defe68cc4dc4bffb53b9cc477f721f5")
+        fetch(fetchUrl)
         .then(result => result.json())
         .then(featureCollection =>{
             //console.log(featureCollection)
@@ -15,6 +19,7 @@ export async function buscador_gasolineras(radio, coords) {
         .catch(error => console.log('error', error))
         );
     }
+
     return Promise.all(fetches).then(function () {
         return setGasolineras
         
@@ -55,26 +60,3 @@ export function buscadorInformacionGasolinera(gasolineras){
     
 
 }
-
-export async function mostrarGasolinerasEnRadio(radio, ubicacionCoords) {
-    // Obtener las gasolineras cercanas utilizando la API de geoapify
-    const response = await fetch(`https://api.geoapify.com/v2/places?categories=service.vehicle.fuel&filter=circle:${ubicacionCoords[1]},${ubicacionCoords[0]},${radio}&limit=20&apiKey=5defe68cc4dc4bffb53b9cc477f721f5`);
-    const data = await response.json();
-  
-    // Crear un array de objetos con las coordenadas y la información relevante de cada gasolinera
-    const gasolinerasEnRadio = [];
-    data.features.forEach(function(gasolinera) {
-      const gasolineraCoords = gasolinera.geometry.coordinates;
-      const distancia = L.latLng(ubicacionCoords).distanceTo(gasolineraCoords.reverse());
-      if (distancia <= radio) {
-        gasolinerasEnRadio.push({
-          latlng: gasolineraCoords.reverse(),
-          name: gasolinera.properties.name,
-          address: gasolinera.properties.formatted_address
-        });
-      }
-    });
-    
-    // Devolver el array de objetos con las coordenadas y la información relevante de cada gasolinera
-    return gasolinerasEnRadio;
-  }
