+++ script.js
@@ -1,7 +1,7 @@
 import { buscadorInformacionGasolinera } from "./gasolineras.js";
 import { buscador_gasolineras } from "./gasolineras.js";
 import { presentadorGasolineras } from "./gasolineras.js";
-import { mostrarGasolinerasEnRadio } from "./gasolineras.js";
+import { mostrarRatio } from "./radio.js";
 //Icono para gasolineras
 var iconGas = new L.icon({
   iconUrl: 'https://raw.githubusercontent.com/pointhi/leaflet-color-markers/master/img/marker-icon-2x-orange.png',
@@ -124,6 +124,7 @@ document
                                 console.log(coordGasolinera);
                                 markers.push(L.marker([parseFloat(coordGasolinera[1]),parseFloat(coordGasolinera[0])]).addTo(map));
                             });*/
+                    result.forEach(i => console.log("RESULT" + i));
                     presentadorGasolineras(markers, result, map, iconGas);
                     buscadorInformacionGasolinera(result);
                   });
@@ -142,67 +143,7 @@ document
     }
   });
 
-//Variables necesarias para mostrar las gasolineras dentro de un ratio
-var circle = null;
-var marks = L.layerGroup().addTo(map);
-
-//Mostrar gasolineras en ratio
-document.getElementById("ubicacion").addEventListener("keypress", function (e) {
-  if (e.keyCode === 13) {
-    e.preventDefault(); //Prevenir el envio del formulario
-
-    var ubicacion = document.getElementById("ubicacion").value;
-
-    //Obtener las coordenadas de la ubicación utilizando la API de OpenStreetMap
-    var ubicacion_url =
-      "https://nominatim.openstreetmap.org/search.php?q=" +
-      ubicacion +
-      "&format=jsonv2";
-    fetch(ubicacion_url)
-      .then(function (response) {
-        return response.json();
-      })
-      .then(function (data) {
-        console.log(data);
-        var ubicacionCoords = [
-          parseFloat(data[0].lat),
-          parseFloat(data[0].lon),
-        ]; //Coordenadas de la ubicación
-
-        console.log(ubicacionCoords);
-
-        //ELiminar marcadores
-        marks.clearLayers();
-
-        // Actualizar el círculo en el mapa
-        if (circle) {
-          map.removeLayer(circle);
-        }
-        circle = L.circle(ubicacionCoords, {
-          radius: 500, // radio en metros
-          color: "blue", // color del borde
-          fillColor: "#3388ff", // color de relleno
-          fillOpacity: 0.2, // opacidad del relleno
-        }).addTo(map);
-
-        // Llamar a la función para obtener las gasolineras dentro del radio y agregar cada marcador a la capa
-        mostrarGasolinerasEnRadio(500, ubicacionCoords).then(function (
-          gasolineras
-        ) {
-          for (let i = 0; i < gasolineras.length; i++) {
-            const gasolinera = gasolineras[i];
-            console.log(gasolinera);
-            marks.addLayer(
-              L.marker([
-                parseFloat(gasolinera.latlng[1]).toFixed(3),
-                parseFloat(gasolinera.latlng[0]).toFixed(3)
-                ],{icon: iconGas}).addTo(map)
-            );
-          }
-        });
-      });
-  }
-});
+mostrarRatio(map, markers, iconGas);
 
 //Eliminar marcadores al cambiar de pagina
 document.getElementById("volver").addEventListener("click", function (e){
